Remove dependency on seelog.
Add a default implementation of the Logger interface so callers do not
need to write their own or create subsystems from seelog loggers.

The Logger interface has been cleaned up to remove error returns from
log methods and to remove the ability to close a logger.  Syncing and
closing must be handled by the application and not packages using the
Logger interface.

The output of the default logger is identical to the current log
output used in dcrd with the exception that time and date have been
reversed.  For btcsuite projects like btcd and btcwallet, this also
means that the logged time also now will include millisecond
resolution (this change was made to Decred projects a short while
ago).

The package API has been modified significantly for main packages that
must setup logging backends, create subsystem loggers, and flushing
when exiting, but only minimal changes have been made to the Logger
interface itself and there should be nothing that requires packages
importing it to be rewritten.