Add flags mechanism and callsite options.
Flags have been added to include the file name/path and line number of
the log callsite in all log messages.

Flags can be configured using the LOGFLAGS environment variable as
well as explicitly setting the flags on a backend in code using
WithFlags.

Inspired by the Lshortfile/Llongfile flags for the stdlib log package.