log: add benchmark tests for V1 and V2 loggers
This commit adds two benchmark tests:

1) BenchmarkLogger tests the new V2 logger under various scenarios.
2) BenchmarkV1vsV2Loggers tests the performance of the V2 logger against
   that of the V1 logger.

As of this commit, the biggest issue is the time it takes for the new v2
logger to perform a log action that actually gets skipped due to the
current log level. This is due to a bug that builds the log string using
fmt.Sprintf _before_ the logger's Enabled check is run against the
current log level. This will be improved upon in the upcoming commits.

The BenchmarkLogger test will also help to assess the various
improvements that will be make in the upcoming commits.

We ideally want BenchmarkV2vsV2Loggers to show that the performance
between the same methods on the two loggers does not change much. This
will be improved upon in upcoming commits.