log: check level before building strings
In this commit, we start checking the log level of the logger's handler
before we do any more evaluation. This does not change behaviour since
the level is checked by the slog library itself later on but by then we
have performed some evaluation which impacts performance.

In the benchmark tests, the biggest differences should be noticeable in
the "skipped" tests.